Timestep + Timescale
Timescale is also passed in to the update loop alongside Timestep. It is combined for the physics and lifetime updates and will be added to the AI and native scripting passes as well. At 1.0f the game runs at standard speed, lower values will allow it to run in "slow motion" as it were, allowing for time adjusting through scripting.